newebuild: Modernize the Python template
Use PEP517 build.  It defaults to setuptools but the eclass will detect
if this is incorrect.

Add a default pypi HOMEPAGE based on the package name.

Skip DEPEND.  Python packages rarely need this class of dependencies,
RDEPEND + BDEPEND are much more common.

Add empty "test?" block in BDEPEND, as this is often necessary.
If it is not, it is easy to yank it.

Add distutils_enable_tests.  It defaults to pytest, as the most common
variant.  It will also trigger explicit failure if the package has
no tests.

Signed-off-by: Michał Górny <mgorny@gentoo.org>