newebuild: Add MAVEN_ID to dev-java ebuild template
Java ebuilds in Gentoo often contain a MAVEN_ID variable.  Although no
eclasses use this variable yet, some Java ebuild tools, like
java-ebuilder, recognize this variable and use its value to establish an
ebuild-to-Maven artifact mapping.  Human maintainers may also use 'grep'
to search for an ebuild with a certain MAVEN_ID for finding out if an
ebuild for a certain Maven artifact already exists.

The MAVEN_ID variable in the dev-java ebuild template acts as a reminder
for Java ebuild authors to record the Maven artifact for the ebuild they
create when applicable.

Signed-off-by: Yuan Liao <liaoyuan@gmail.com>