newebuild: Do not use *.zip SRC_URI in dev-java ebuild template
Many Java packages' source archives are available in tarball formats
like .tar.gz already, and tarballs are more preferable than Zip archives
for SRC_URI (pkgcheck has a check for this).

However, instead of replacing '.zip' with '.tar.gz', it could be better
to just leave the value of SRC_URI in the template empty, just like in
the dev-python ebuild template.

Signed-off-by: Yuan Liao <liaoyuan@gmail.com>
Closes: https://github.com/gentoo/gentoo-syntax/pull/58
Signed-off-by: Michał Górny <mgorny@gentoo.org>